Change the license to use the new endpoint and format (#15091)
* Change the license to use the new endpoint and format

1. Use the `_license` endpoint instead of using the `/_xpack`
2. Remove the `features` handling.
3. Add more golden files for the different returned format.
4. remove unused files and code

(cherry picked from commit 556e8ff566856efc96b4c6ad86069e8aadc1c30e)
