+++ x-pack/libbeat/licenser/check_test.go
@@ -24,7 +24,7 @@ func testCheckTrial(t *testing.T) {
 
 	t.Run("valid trial license", func(t *testing.T) {
 		l := License{
-			Mode:        Trial,
+			Type:        Trial,
 			TrialExpiry: expiryTime(time.Now().Add(1 * time.Hour)),
 		}
 		assert.True(t, CheckTrial(log, l))
@@ -32,14 +32,14 @@ func testCheckTrial(t *testing.T) {
 
 	t.Run("expired trial license", func(t *testing.T) {
 		l := License{
-			Mode:        Trial,
+			Type:        Trial,
 			TrialExpiry: expiryTime(time.Now().Add(-1 * time.Hour)),
 		}
 		assert.False(t, CheckTrial(log, l))
 	})
 
 	t.Run("other license", func(t *testing.T) {
-		l := License{Mode: Basic}
+		l := License{Type: Basic}
 		assert.False(t, CheckTrial(log, l))
 	})
 }
@@ -51,19 +51,19 @@ func testCheckLicenseCover(t *testing.T) {
 		fn := CheckLicenseCover(license)
 
 		t.Run("active", func(t *testing.T) {
-			l := License{Mode: license, Status: Active}
+			l := License{Type: license, Status: Active}
 			assert.True(t, fn(log, l))
 		})
 
 		t.Run("inactive", func(t *testing.T) {
-			l := License{Mode: license, Status: Inactive}
+			l := License{Type: license, Status: Inactive}
 			assert.False(t, fn(log, l))
 		})
 	}
 }
 
 func testValidate(t *testing.T) {
-	l := License{Mode: Basic, Status: Active}
+	l := License{Type: Basic, Status: Active}
 	t.Run("when one of the check is valid", func(t *testing.T) {
 		valid := Validate(logp.NewLogger(""), l, CheckLicenseCover(Platinum), CheckLicenseCover(Basic))
 		assert.True(t, valid)
