--- x-pack/libbeat/licenser/testdata/xpack-6.4.0.json
@@ -1,43 +1,0 @@
-{
-  "build": {
-    "hash": "053779d",
-    "date": "2018-07-20T05:25:16.206115Z"
-  },
-  "license": {
-    "uid": "936183d8-f48c-4a3f-959a-a52aa2563279",
-    "type": "platinum",
-    "mode": "platinum",
-    "status": "active",
-    "expiry_date_in_millis": 1588291199999
-  },
-  "features": {
-    "graph": {
-      "available": false,
-      "enabled": true
-    },
-    "logstash": {
-      "available": false,
-      "enabled": true
-    },
-    "ml": {
-      "available": false,
-      "enabled": true
-    },
-    "monitoring": {
-      "available": true,
-      "enabled": true
-    },
-    "rollup": {
-      "available": true,
-      "enabled": true
-    },
-    "security": {
-      "available": false,
-      "enabled": true
-    },
-    "watcher": {
-      "available": false,
-      "enabled": true
-    }
-  }
-}
