--- x-pack/libbeat/licenser/testdata/xpack-trial-6.4.0.json
@@ -1,43 +1,0 @@
-{
-  "build": {
-    "hash": "595516e",
-    "date": "2018-08-17T23:22:27.102119Z"
-  },
-  "license": {
-    "uid": "936183d8-f48c-4a3f-959a-a52aa2563279",
-    "type": "trial",
-    "mode": "trial",
-    "status": "active",
-    "expiry_date_in_millis": 1588291199999
-  },
-  "features": {
-    "graph": {
-      "available": false,
-      "enabled": true
-    },
-    "logstash": {
-      "available": false,
-      "enabled": true
-    },
-    "ml": {
-      "available": false,
-      "enabled": true
-    },
-    "monitoring": {
-      "available": true,
-      "enabled": true
-    },
-    "rollup": {
-      "available": true,
-      "enabled": true
-    },
-    "security": {
-      "available": false,
-      "enabled": true
-    },
-    "watcher": {
-      "available": false,
-      "enabled": true
-    }
-  }
-}
