+++ x-pack/libbeat/licenser/testdata/xpack-with-standard.json
@@ -5,7 +5,7 @@
   },
   "license": {
     "uid": "936183d8-f48c-4a3f-959a-a52aa2563279",
-    "type": "enterprise",
+    "type": "standard",
     "issue_date_in_millis": 1576000000000,
     "start_date_in_millis": 1576000000000,
     "expiry_date_in_millis": 1588291199999,
